Updated per code review
Added Converters = { new JsonStringEnumConverter(JsonNamingPolicy.CamelCase) } in JsonSeriralization options for the 'sort' ListRecords parameter.
Removed SortWithDirectinString.
Moved UserIdAndScopes from QueryParamHelper.cs to new file UserIdAndScopes.cs.
